--- frontend/image_upload/.angular/cache/19.1.6/image_upload/vite/deps_ssr/@angular_animations.js
@@ -1,50 +1,0 @@
-import { createRequire } from 'module';const require = createRequire(import.meta.url);
-import {
-  AUTO_STYLE,
-  AnimationBuilder,
-  AnimationFactory,
-  AnimationGroupPlayer,
-  AnimationMetadataType,
-  BrowserAnimationBuilder,
-  NoopAnimationPlayer,
-  animate,
-  animateChild,
-  animation,
-  group,
-  keyframes,
-  query,
-  sequence,
-  stagger,
-  state,
-  style,
-  transition,
-  trigger,
-  useAnimation,
-  ɵPRE_STYLE
-} from "./chunk-UJA52W4A.js";
-import "./chunk-YAGPW5AS.js";
-import "./chunk-53Z7NWHP.js";
-import "./chunk-YHCV7DAQ.js";
-export {
-  AUTO_STYLE,
-  AnimationBuilder,
-  AnimationFactory,
-  AnimationMetadataType,
-  NoopAnimationPlayer,
-  animate,
-  animateChild,
-  animation,
-  group,
-  keyframes,
-  query,
-  sequence,
-  stagger,
-  state,
-  style,
-  transition,
-  trigger,
-  useAnimation,
-  AnimationGroupPlayer as ɵAnimationGroupPlayer,
-  BrowserAnimationBuilder as ɵBrowserAnimationBuilder,
-  ɵPRE_STYLE
-};
