+++ frontend/image_upload/.angular/cache/19.1.6/image_upload/vite/deps_ssr/@angular_platform-server.js
@@ -13,8 +13,7 @@ import {
   provideServerRendering,
   renderApplication,
   renderModule
-} from "./chunk-BIXN34VP.js";
-import "./chunk-UJA52W4A.js";
+} from "./chunk-WGHIGEYG.js";
 import "./chunk-FZHFUH5J.js";
 import "./chunk-RSDZCMNU.js";
 import "./chunk-YAGPW5AS.js";
