+++ frontend/image_upload/.angular/cache/19.1.6/image_upload/vite/deps_ssr/@angular_ssr.js
@@ -14,10 +14,9 @@ import {
   setAngularAppEngineManifest,
   setAngularAppManifest,
   withAppShell
-} from "./chunk-VD2GKPCJ.js";
-import "./chunk-BIXN34VP.js";
-import "./chunk-UJA52W4A.js";
+} from "./chunk-O3M6X5EB.js";
 import "./chunk-7K2GJIEI.js";
+import "./chunk-WGHIGEYG.js";
 import "./chunk-FZHFUH5J.js";
 import "./chunk-RSDZCMNU.js";
 import "./chunk-YAGPW5AS.js";
