+++ frontend/image_upload/.angular/cache/19.1.6/image_upload/vite/deps_ssr/@angular_ssr_node.js
@@ -2,14 +2,13 @@ import { createRequire } from 'module';const require = createRequire(import.meta
 import {
   AngularAppEngine,
   InlineCriticalCssProcessor
-} from "./chunk-VD2GKPCJ.js";
+} from "./chunk-O3M6X5EB.js";
+import "./chunk-7K2GJIEI.js";
 import {
   SERVER_CONTEXT,
   renderApplication,
   renderModule
-} from "./chunk-BIXN34VP.js";
-import "./chunk-UJA52W4A.js";
-import "./chunk-7K2GJIEI.js";
+} from "./chunk-WGHIGEYG.js";
 import "./chunk-FZHFUH5J.js";
 import "./chunk-RSDZCMNU.js";
 import "./chunk-YAGPW5AS.js";
