+++ frontend/image_upload/.angular/cache/19.1.6/image_upload/vite/deps_ssr/_metadata.json
@@ -1,85 +1,76 @@
 {
-  "hash": "c44c73d4",
-  "configHash": "fb8e0817",
+  "hash": "7bee628f",
+  "configHash": "ec17b61f",
   "lockfileHash": "e8750ed7",
-  "browserHash": "31a6417b",
+  "browserHash": "a3c59be5",
   "optimized": {
     "@angular/common": {
       "src": "../../../../../../node_modules/@angular/common/fesm2022/common.mjs",
       "file": "@angular_common.js",
-      "fileHash": "da740e68",
+      "fileHash": "2debe516",
       "needsInterop": false
     },
     "@angular/common/http": {
       "src": "../../../../../../node_modules/@angular/common/fesm2022/http.mjs",
       "file": "@angular_common_http.js",
-      "fileHash": "102512d1",
+      "fileHash": "ede87dc6",
       "needsInterop": false
     },
     "@angular/core": {
       "src": "../../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "66faa875",
+      "fileHash": "973dbc0d",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "d3557955",
+      "fileHash": "c2456e92",
       "needsInterop": false
     },
     "@angular/platform-server": {
       "src": "../../../../../../node_modules/@angular/platform-server/fesm2022/platform-server.mjs",
       "file": "@angular_platform-server.js",
-      "fileHash": "262ce5cf",
+      "fileHash": "3ace05de",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "dd04eff4",
+      "fileHash": "e97b702e",
       "needsInterop": false
     },
     "@angular/ssr": {
       "src": "../../../../../../node_modules/@angular/ssr/fesm2022/ssr.mjs",
       "file": "@angular_ssr.js",
-      "fileHash": "2352f99e",
+      "fileHash": "6e0ab563",
       "needsInterop": false
     },
     "@angular/ssr/node": {
       "src": "../../../../../../node_modules/@angular/ssr/fesm2022/node.mjs",
       "file": "@angular_ssr_node.js",
-      "fileHash": "fbc7b3c0",
+      "fileHash": "c9cca3d7",
       "needsInterop": false
     },
     "express": {
       "src": "../../../../../../node_modules/express/index.js",
       "file": "express.js",
-      "fileHash": "4574bb9b",
+      "fileHash": "faad7619",
       "needsInterop": true
-    },
-    "@angular/animations": {
-      "src": "../../../../../../node_modules/@angular/animations/fesm2022/animations.mjs",
-      "file": "@angular_animations.js",
-      "fileHash": "bfe5224c",
-      "needsInterop": false
     }
   },
   "chunks": {
     "xhr2-TXIMV6CV": {
       "file": "xhr2-TXIMV6CV.js"
     },
-    "chunk-VD2GKPCJ": {
-      "file": "chunk-VD2GKPCJ.js"
-    },
-    "chunk-BIXN34VP": {
-      "file": "chunk-BIXN34VP.js"
-    },
-    "chunk-UJA52W4A": {
-      "file": "chunk-UJA52W4A.js"
+    "chunk-O3M6X5EB": {
+      "file": "chunk-O3M6X5EB.js"
     },
     "chunk-7K2GJIEI": {
       "file": "chunk-7K2GJIEI.js"
+    },
+    "chunk-WGHIGEYG": {
+      "file": "chunk-WGHIGEYG.js"
     },
     "chunk-FZHFUH5J": {
       "file": "chunk-FZHFUH5J.js"
