+++ frontend/image_upload/src/app/components/upload-form/upload-form.component.html
@@ -83,6 +83,6 @@
 <div *ngIf='isLoading' class='overlay'>
   <div class='spinner-container'>
     <div class='spinner'></div>
-    <p>Board is being solved…</p>
+    <p>Solving...</p>
   </div>
 </div>
