+++ frontend/image_upload/src/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
 <head>
   <meta charset="utf-8">
-  <title>ImageUpload</title>
+  <title>Block Blast Solver</title>
   <base href="/">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" type="image/x-icon" href="favicon.ico">
+  <link rel="icon" type="image/jpeg" href="https://product-list.sfo3.digitaloceanspaces.com/products/block-blast-solver-6/images/a981416b-596c-4d95-88eb-d0c6823a974f.jpeg">
 </head>
 <script src="https://unpkg.com/@phosphor-icons/web"></script>
 <body>
